SLING-11340 cleanup sonar problems
consolidate duplicate code block to a common base class
